requirements/dev.txt
-r defaults.txt
django-nose==1.4.5
coverage==4.5.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
Pillow==5.2.0
